パッケージの追加とクラス等の整理、新たにクラスを追加しRead処理のServiceとController部分を実装
